# openapi.yaml
openapi: 3.1.0
info:
  title: Bulk Food by QML API
  version: 0.1.0
  summary: Backend API for the QML customer, vendor, financier, logistics, corporate, and admin flows.
  description: >
    This API powers marketplace listings, AI approval, role dashboards, wallets,
    orders, KYC, support, refunds, reports, and admin oversight. External APIs
    should be used for identity, payments, maps, notifications, and storage while
    QML owns business logic, risk rules, wallet ledger decisions, and approvals.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://bulk-food-by-qml-api.onrender.com
    description: Render production
security:
  - AdminRoleHeader: []
    AdminApiKey: []
...
paths:
...
  /api/packages:
    get:
      tags: [Marketplace]
      summary: List food packages with optional country/city filtering.
      security: []
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Food packages.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodPackage"
...
  /api/role-applications:
    post:
      tags: [Users]
      summary: Apply for vendor, financier, logistics, or corporate role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId:
                  type: string
                role:
                  type: string
                  enum: [vendor, financier, logistics, corporate]
                referralCode:
                  type: string
                payload:
                  type: object
      responses:
        "201":
          description: Application submitted.
components:
  securitySchemes:
    AdminRoleHeader:
      type: apiKey
      in: header
      name: X-Admin-Role
      description: Development/admin role header. Replace with JWT auth before public production.
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
      description: Optional production hardening header. Required when ADMIN_API_KEY is configured.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT
...
  schemas:
...
    FoodPackage:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        price:
          type: number
        category:
          type: string
        city:
          type: string
...